requirements.txt
python-telegram-bot[job-queue]>=22.6,<23
librouteros>=4.0.0,<5
python-dotenv>=1.0.0,<2
matplotlib>=3.10.0,<3.11

# Testing
pytest>=9.0.0,<10
pytest-asyncio>=1.3.0,<2
